nongnu: google-chrome-stable: Update to 108.0.5359.124.
The main point of this patch is to make Chrome rely as little as
possible on the environment it's run into, allowing for easier
debugging. The next step would probably be to generate this list from
the input package-name.

* nongnu/packages/chrome.scm (google-chrome-stable): Update to 108.0.5359.124;
[arguments]{patchelf-inputs}: move to outer scope;
{LD_LIBRARY_PATH}: wrap the entrypoint with all of Chrome's input, this
should prevent loading libraries that are not listed as inputs from the
outer environment;
[inputs]: add some inputs following Nix's package definition to try and
fix #215.

Signed-off-by: Jonathan Brielmaier <jonathan.brielmaier@web.de>