nongnu: Keep the raw-initrd references file.
The raw-initrd contains a "references" file that is used to keep the static
guile used in the initrd alive. This file is not part of the combined-initrd.

It means that during garbage collection, the static guile could be collected
making the system unbootable because the static guile is then not part of the
store once the root is switched.

In the combined-initrds procedure, make sure to concatenate all the possible
references files of the underlying initrds into a top-level references file.

Fixes: <https://gitlab.com/nonguix/nonguix/issues/111>
Signed-off-by: Jonathan Brielmaier <jonathan.brielmaier@web.de>