nonguix: chromium-binary-build-system: Use gcc:lib.
Effectively reverts changes that came from
9d58bb6e3e93a444d35bf62d4ca98a126ae4f295 which doesn't seem to be needed
anymore.  These got carried into chromium-binary-build-system when it was
introduced.  We need libstdc++ from the gcc:lib package, which is hidden from
users but directly accessible from the build system.  The library from
make-libstdc++ has missing symbols compared to the one from gcc:lib.  The
addition of libgccjit seems to also be unneeded from this change and increases
the closure of e.g. signal-desktop by nearly 10% or 200 MiB.

Fixes #276 and #279.

* nonguix/build-system/chromium-binary.scm (lower)[host-inputs]: Use gcc:lib
again and remove libgccjit.