nongnu: linux: Enable mt7921e driver.
In order to be able to use WiFi chip in P14s ThinkPad, the mt7921e module is
required.  This commit enables it in kernels that do have it.

I originally implemented this using a list of additional option, same way
%default-extra-linux-options is done.  However I quickly realized that
approach is not suitable for hardware enablement.  The older versions do not
support the same drivers as the newer ones.

Solution is to create a new procedure that generates the list for a specific
kernel version and use it as a default value for #:configs.

* nongnu/packages/linux.scm (nonguix-extra-linux-options): New procedure.
(corrupt-linux): Use it as default value for #:configs.

Signed-off-by: Jonathan Brielmaier <jonathan.brielmaier@web.de>