nongnu: corrupt-linux: Pass 'linux' argument to customize-linux.
This fixes an issue where corrupt-linux would use the kernel config from the
wrong version.  The reason is that the package phases in customize-linux come
from inheriting the package in the 'linux' keyword argument, defaulting to
linux-libre.  The kernel configuration comes in the configure phase.  Since we
did not use this in the customize-linux in corrupt-linux, we would be using
the wrong configuration.  For example, linux-6.7 would be built with the
linux-libre-6.6 kernel config rather than linux-libre-6.7, missing the new
CONFIG_BCACHEFS_FS option.  Presumably all of our kernels were trying to use
the default linux-libre version kernel config rather than the correct one to
match their version.  This should fix that.

* nongnu/packages/linux.scm (corrupt-linux): Use the 'linux' keyword argument
in customize-linux.

Reported-by: reedm and yelninei on the #nonguix IRC channel.