nongnu: Remove icu4c-73.
This version of icu4c is in Guix and made public in
cf842e9b20b89ecb08ac3456a91780ec07b5a201.  This also failed to build on the
core-updates nonguix build due to a test failure which should be fixed in Guix
with 7937c8827b8d23347a3159b4696335bd19fc17aa.

* nongnu/packages/mozilla.scm (icu4c-73): Delete variable.