nongnu: nvidia-nvml: Fix baked-in source in unpack overrides.
* nongnu/packages/nvidia.scm (nvidia-nvml)[#:phases]<unpack>: Replace source
G-expression by keyword argument source.

These cases of `replace 'unpack` were referring to the `#$source` of the
package in the g-exp, which was baking in the source *at the point of
package definition* to the g-exp. This prevented inheritor packages from
modifying the source field and getting the expected effect.

The correct way to refer to the source field from a phase override is
with the `source` key. This modifies all the references to `#$source` I
could find in `unpack` phase replacements.

Signed-off-by: Jelle Licht <jlicht@fsfe.org>