nonguix: multiarch-container: Don't try to start pulseaudio.
Closes #241.

See discussion in above issue, but the summary is that the container (or any
package/program) shouldn't be trying to start pulseaudio (or other home/system
services, especially that can be started on demand).  While previously this
was done to avoid some issue, it is not clear if that is still the case and it
may possibly lead to other issues.  So, let's default to a more
expected (non-) behavior.

* nonguix/multiarch-container.scm (make-container-wrapper): Remove invoke of
pulseaudio.