nongnu: steam-client-libs: Add fonts for emojis and CJK scripts.
The choice of fonts here is arbitrary, chosen as the google-noto set has fonts
for emoji and Chinese/Japanese/Korean and related scripts.  This fixes missing
emoji and characters from these languages in the Steam interface.  It would be
better to be able to use the font configuration and fonts from the host system
in the container, though.

* nongnu/packages/game-client.scm (steam-client-libs): Add font-google-noto,
font-google-noto-emoji, font-google-noto-sans-cjk, font-google-noto-serif-cjk.